Improve detection of sitecustomize.pl
Debian Perl is shipped with the following patch, which loads sitecustomize.pl
from a non-standard location:

    DEBPKG:debian/CVE-2016-1238/sitecustomize-in-etc - Look for sitecustomize.pl in /etc/perl rather than sitelib on Debian systems

This change is enough to fix the test failure under Debian, but the code
around this seems too complicated for what it actually does. Picking the
location of sitecustomize.pl from %INC should be sufficient.