Consider each attempt at loading a module as an unique thing
The $TRACE{used} hash was conflating successive failed attempts to load
a given module into a single data structure. If after several failed
attempts at loading a module there is a success, all the attemps would
be showed using the latest data (from the success).

The $TRACE{ranked} array keeps a different data structure for each
attempt, and the loaded_by keys in those structures actually points to
the individual items in the array, thus making it possible to point at
the actual individual attempt.